frontend mostly done
