+++ client/eslint.config.js
@@ -11,6 +11,10 @@ export default tseslint.config({
   languageOptions: {
     ecmaVersion: 2020,
     globals: globals.browser,
+    parserOptions: {
+      project: "./tsconfig.app.json",
+      tsconfigRootDir: import.meta.dirname,
+    },
   },
   plugins: {
     "react-hooks": reactHooks,
