+++ client/src/components/login/Login.tsx
@@ -1,179 +1,273 @@
-import { useCallback, useEffect } from "react";
-
-import {
-  Button,
-  Center,
-  Field,
-  Link as ChakraLink,
-  Heading,
-  Input,
-  Stack,
-  VStack,
-} from "@chakra-ui/react";
-
-import { zodResolver } from "@hookform/resolvers/zod";
-import { useForm } from "react-hook-form";
-import { FaGoogle } from "react-icons/fa";
-import { Link, useNavigate } from "react-router-dom";
-import { z } from "zod";
-
-import { toaster } from "@/components/ui/toaster";
-import { useAuthContext } from "@/contexts/hooks/useAuthContext";
-import { useBackendContext } from "@/contexts/hooks/useBackendContext";
-import { authenticateGoogleUser } from "@/utils/auth/providers";
-
-const signinSchema = z.object({
-  email: z.string().email("Invalid email address"),
-  password: z.string().min(6, "Password must be at least 6 characters long"),
-});
-
-type SigninFormValues = z.infer<typeof signinSchema>;
+import { Box, Button, Flex, Heading, Text, Link, Icon, Stack } from "@chakra-ui/react";
+import { FaBriefcase, FaUser, FaFacebook, FaLinkedin, FaInstagram, FaEnvelope, FaArrowRight } from "react-icons/fa";
+import { useNavigate } from "react-router-dom";
 
 export const Login = () => {
   const navigate = useNavigate();
 
-  const { login, handleRedirectResult } = useAuthContext();
-  const { backend } = useBackendContext();
-
-  const {
-    register,
-    handleSubmit,
-    formState: { errors },
-  } = useForm<SigninFormValues>({
-    resolver: zodResolver(signinSchema),
-    mode: "onBlur",
-  });
-
-  const toastLoginError = useCallback(
-    (msg: string) => {
-      toaster.error({
-        title: "An error occurred while signing in",
-        description: msg,
-      });
-    },
-    []
-  );
-
-  const handleLogin = async (data: SigninFormValues) => {
-    try {
-      await login({
-        email: data.email,
-        password: data.password,
-      });
-
-      navigate("/dashboard");
-    } catch (err) {
-      const errorCode = err.code;
-      const firebaseErrorMsg = err.message;
-
-      switch (errorCode) {
-        case "auth/wrong-password":
-        case "auth/invalid-credential":
-        case "auth/invalid-email":
-        case "auth/user-not-found":
-          toastLoginError(
-            "Email address or password does not match our records!"
-          );
-          break;
-        case "auth/unverified-email":
-          toastLoginError("Please verify your email address.");
-          break;
-        case "auth/user-disabled":
-          toastLoginError("This account has been disabled.");
-          break;
-        case "auth/too-many-requests":
-          toastLoginError("Too many attempts. Please try again later.");
-          break;
-        case "auth/user-signed-out":
-          toastLoginError("You have been signed out. Please sign in again.");
-          break;
-        default:
-          toastLoginError(firebaseErrorMsg);
-      }
-    }
+  const handleVolunteerClick = () => {
+    navigate("/login/volunteer");
   };
 
-  const handleGoogleLogin = async () => {
-    await authenticateGoogleUser();
+  const handleStaffClick = () => {
+    navigate("/login/staff");
   };
-
-  useEffect(() => {
-    handleRedirectResult(backend, navigate);
-  }, [backend, handleRedirectResult, navigate]);
 
   return (
-    <VStack
-      gap={8}
-      w="250px"
-      mx="auto"
-    >
-      <Heading>Login</Heading>
-
-      <form
-        onSubmit={handleSubmit(handleLogin)}
-        style={{ width: "100%" }}
-      >
-        <Stack gap={2}>
-          <Field.Root
-            invalid={!!errors.email}
-            w={"100%"}
-          >
-            <Center>
-              <Input
-                placeholder="Email"
-                type="email"
-                size={"lg"}
-                {...register("email")}
-                name="email"
-                required
-                autoComplete="email"
-              />
-            </Center>
-            <Field.ErrorText>
-              {errors.email?.message?.toString()}
-            </Field.ErrorText>
-          </Field.Root>
-          <Field.Root invalid={!!errors.password}>
-            <Center>
-              <Input
-                placeholder="Password"
-                type="password"
-                size={"lg"}
-                {...register("password")}
-                name="password"
-                required
-                autoComplete="current-password"
-              />
-            </Center>
-            <Field.ErrorText>
-              {errors.password?.message?.toString()}
-            </Field.ErrorText>
-            <ChakraLink asChild>
-              <Link to="/signup">
-              <Field.HelperText>Click here to sign up</Field.HelperText>
-              </Link>
-            </ChakraLink>
-          </Field.Root>
-
-          <Button
-            type="submit"
-            size={"lg"}
-            w="100%"
-            disabled={Object.keys(errors).length > 0}
-          >
-            Login
-          </Button>
-        </Stack>
-      </form>
-
-      <Button
-        variant="solid"
-        size="lg"
-        onClick={handleGoogleLogin}
+    <>
+      {/* Mobile Layout */}
+      <Flex
+        display={{ base: "flex", md: "none" }}
         w="100%"
+        minH="100vh"
+        direction="column"
+        bg="white"
+        px={6}
+        py={8}
+        alignItems="center"
+        justifyContent="space-between"
       >
-        Login with Google <FaGoogle />
-      </Button>
-    </VStack>
+        {/* Main Content */}
+        <Box w="full" maxW="400px" flex="1" display="flex" flexDirection="column" justifyContent="center">
+          {/* Header */}
+          <Heading
+            fontSize="2xl"
+            fontWeight="bold"
+            lineHeight="1.3"
+            mb={8}
+            color="black"
+          >
+            CC Login Portal by Community Counsel
+          </Heading>
+
+          {/* Welcome Section */}
+          <Heading
+            fontSize="xl"
+            fontWeight="bold"
+            mb={3}
+            color="black"
+          >
+            Welcome!
+          </Heading>
+          <Text
+            fontSize="sm"
+            color="gray.500"
+            mb={8}
+          >
+            Choose from the options below to continue.
+          </Text>
+
+          {/* Buttons */}
+          <Stack gap={4} w="full" mb={12}>
+            <Button
+              bg="blue.500"
+              color="white"
+              w="full"
+              h="56px"
+              fontSize="md"
+              fontWeight="semibold"
+              borderRadius="md"
+              _hover={{ bg: "blue.600" }}
+              display="flex"
+              justifyContent="space-between"
+              alignItems="center"
+              px={6}
+              onClick={handleStaffClick}
+            >
+              <Icon as={FaBriefcase} boxSize={5} />
+              <Text flex="1" textAlign="center">
+                Staff Member
+              </Text>
+              <Icon as={FaArrowRight} boxSize={5} />
+            </Button>
+            <Button
+              bg="blue.500"
+              color="white"
+              w="full"
+              h="56px"
+              fontSize="md"
+              fontWeight="semibold"
+              borderRadius="md"
+              _hover={{ bg: "blue.600" }}
+              display="flex"
+              justifyContent="space-between"
+              alignItems="center"
+              px={6}
+              onClick={handleVolunteerClick}
+            >
+              <Icon as={FaUser} boxSize={5} />
+              <Text flex="1" textAlign="center">
+                Volunteer
+              </Text>
+              <Icon as={FaArrowRight} boxSize={5} />
+            </Button>
+          </Stack>
+
+          {/* Help Section */}
+          <Box mb={6}>
+            <Text fontSize="md" fontWeight="bold" mb={0}>
+              Need help?
+            </Text>
+            <Text fontSize="md" fontWeight="bold">
+              Visit our website
+            </Text>
+          </Box>
+          <Link
+            href="https://communitycounsel.org"
+            color="blue.500"
+            fontSize="md"
+            textDecoration="underline"
+            mb={8}
+            display="block"
+          >
+            Community Counsel Website
+          </Link>
+        </Box>
+
+        {/* Footer - Social Icons */}
+        <Flex gap={6} justifyContent="center" w="full" py={4}>
+          <Icon as={FaFacebook} boxSize={6} color="black" cursor="pointer" />
+          <Icon as={FaLinkedin} boxSize={6} color="black" cursor="pointer" />
+          <Icon as={FaInstagram} boxSize={6} color="black" cursor="pointer" />
+          <Icon as={FaEnvelope} boxSize={6} color="black" cursor="pointer" />
+        </Flex>
+      </Flex>
+
+      {/* Desktop Layout */}
+      <Flex
+        display={{ base: "none", md: "flex" }}
+        w="100%"
+        h="100vh"
+        minH="600px"
+      >
+        {/* Left Side */}
+        <Flex
+          w="50%"
+          bg="white"
+          direction="column"
+          justifyContent="space-between"
+          px={12}
+          py={16}
+        >
+          <Box>
+            <Heading
+              fontSize={{ base: "3xl", lg: "4xl", xl: "5xl" }}
+              fontWeight="bold"
+              lineHeight="1.2"
+              mb={6}
+              color="black"
+            >
+              Welcome to CC Login Portal by Community Counsel
+            </Heading>
+            <Text
+              fontSize={{ base: "lg", lg: "xl" }}
+              color="gray.700"
+            >
+              Indicate whether you are a staff member or volunteer
+            </Text>
+          </Box>
+
+          <Box>
+            <Box mb={6}>
+              <Text fontSize="xl" fontWeight="bold" mb={2}>
+                Need help?
+              </Text>
+              <Text fontSize="xl" fontWeight="bold">
+                Visit our website
+              </Text>
+            </Box>
+            <Link
+              href="https://communitycounsel.org"
+              color="blue.500"
+              fontSize="lg"
+              textDecoration="underline"
+              mb={8}
+              display="block"
+            >
+              Community Counsel Website
+            </Link>
+            <Flex gap={4}>
+              <Icon as={FaFacebook} boxSize={6} color="black" cursor="pointer" />
+              <Icon as={FaLinkedin} boxSize={6} color="black" cursor="pointer" />
+              <Icon as={FaInstagram} boxSize={6} color="black" cursor="pointer" />
+              <Icon as={FaEnvelope} boxSize={6} color="black" cursor="pointer" />
+            </Flex>
+          </Box>
+        </Flex>
+
+        {/* Right Side */}
+        <Flex
+          flex="1"
+          bg="gray.50"
+          direction="column"
+          alignItems="center"
+          justifyContent="flex-start"
+          px={12}
+          pt={48}
+        >
+          <Box w="full" maxW="500px">
+            <Heading
+              fontSize="4xl"
+              fontWeight="bold"
+              mb={3}
+              color="black"
+            >
+              Welcome!
+            </Heading>
+            <Text
+              fontSize="md"
+              color="gray.500"
+              mb={10}
+            >
+              Choose from the options below to continue.
+            </Text>
+            <Stack gap={4} w="full">
+              <Button
+                bg="blue.500"
+                color="white"
+                w="full"
+                h="50px"
+                fontSize="lg"
+                fontWeight="semibold"
+                borderRadius="sm"
+                _hover={{ bg: "blue.600" }}
+                display="flex"
+                justifyContent="space-between"
+                alignItems="center"
+                px={6}
+                onClick={handleStaffClick}
+              >
+                <Icon as={FaBriefcase} boxSize={5} />
+                <Text flex="1" textAlign="center">
+                  Staff Member
+                </Text>
+                <Icon as={FaArrowRight} boxSize={5} />
+              </Button>
+              <Button
+                bg="blue.500"
+                color="white"
+                w="full"
+                h="50px"
+                fontSize="lg"
+                fontWeight="semibold"
+                borderRadius="sm"
+                _hover={{ bg: "blue.600" }}
+                display="flex"
+                justifyContent="space-between"
+                alignItems="center"
+                px={6}
+                onClick={handleVolunteerClick}
+              >
+                <Icon as={FaUser} boxSize={5} />
+                <Text flex="1" textAlign="center">
+                  Volunteer
+                </Text>
+                <Icon as={FaArrowRight} boxSize={5} />
+              </Button>
+            </Stack>
+          </Box>
+        </Flex>
+      </Flex>
+    </>
   );
 };
